renames for clarity and correctness
