+++ moveit_experimental/moveit_jog_arm/src/jog_calcs.cpp
@@ -40,7 +40,7 @@
 #include <std_msgs/Bool.h>
 
 #include <moveit_jog_arm/jog_calcs.h>
-#include <moveit_jog_arm/boost_pool_allocation.h>
+#include <moveit_jog_arm/make_shared_from_pool.h>
 
 static const std::string LOGNAME = "jog_calcs";
 
@@ -113,9 +113,9 @@ JogCalcs::JogCalcs(ros::NodeHandle& nh, const JogArmParameters& parameters,
       internal_nh.subscribe("collision_velocity_scale", 1, &JogCalcs::collisionVelocityScaleCB, this);
 
   // Wait for initial messages
-  ROS_INFO_NAMED(LOGNAME, "jog_calcs_thread: Waiting for first joint msg.");
+  ROS_INFO_NAMED(LOGNAME, "Waiting for first joint msg.");
   ros::topic::waitForMessage<sensor_msgs::JointState>(parameters_.joint_topic);
-  ROS_INFO_NAMED(LOGNAME, "jog_calcs_thread: Received first joint msg.");
+  ROS_INFO_NAMED(LOGNAME, "Received first joint msg.");
 
   internal_joint_state_.name = joint_model_group_->getActiveJointModelNames();
   num_joints_ = internal_joint_state_.name.size();
@@ -163,7 +163,7 @@ void JogCalcs::run(const ros::TimerEvent& timer_event)
 
   // Publish status each loop iteration
   {
-    auto status_msg = make_shared_from_pool<std_msgs::Int8>();
+    auto status_msg = moveit::util::make_shared_from_pool<std_msgs::Int8>();
     status_msg->data = static_cast<int8_t>(status_);
     status_pub_.publish(status_msg);
   }
@@ -221,7 +221,7 @@ void JogCalcs::run(const ros::TimerEvent& timer_event)
   else
   {
     // Create new outgoing joint trajectory command message
-    auto joint_trajectory = make_shared_from_pool<trajectory_msgs::JointTrajectory>();
+    auto joint_trajectory = moveit::util::make_shared_from_pool<trajectory_msgs::JointTrajectory>();
 
     // If the input command isn't stale, run calculations
     if (!command_is_stale_)
